Создан примитивный интерфейс построение множества Жюлиа, но пока не рабочий. Разбил некоторый код для удобства на функции
