+++ src/main/kotlin/math/fractals/AlgebraicFractal.kt
@@ -4,6 +4,8 @@ import math.Complex
 
 interface AlgebraicFractal {
     var maxIterations: Int
+    var function: (Complex) -> Complex
+
     fun isInSet(c: Complex): Float
 
 }
