+++ src/main/kotlin/math/fractals/Fractal.kt
@@ -6,7 +6,12 @@ object Fractal : AlgebraicFractal {
     override var maxIterations: Int = 5000
         set(value) { field = value.coerceIn(20..10000)}
     var r = 2.0
-    var function: (Complex) -> Complex = {value:Complex -> value }
+    override var function: (Complex) -> Complex = { value:Complex -> value }
+        set(value) {
+            if(funcs.containsValue(value)){
+                field = value
+            }
+        }
 
 
     override fun isInSet(c: Complex): Float {
