+++ src/main/kotlin/math/fractals/Julia.kt
@@ -1,19 +1,19 @@
 package math.fractals
+import androidx.compose.ui.geometry.Offset
 import math.Complex
 
+//    companion object {
+//        var selectedPoint =
+//    }
 
-
-class Julia : AlgebraicFractal {
-    companion object {
-        var selectedPoint = Complex(0.0,0.0)
-    }
+class Julia(var selectedPoint: Complex, override var function: (Complex) -> Complex) : AlgebraicFractal {
     override var maxIterations: Int = 2000
         private var r = 2.0
         val r2 = r * r
     override fun isInSet(c : Complex) : Float {
         var z: Complex  = c
         for(i in 0..maxIterations){
-            z = z * z + selectedPoint
+            z = z * z + Complex(selectedPoint.re.toDouble(), selectedPoint.im.toDouble())
             if (z.abs2() >= r2)
                 return i.toFloat() / maxIterations
         }
